yaml-controller/requirements.txt
pyyaml==6.0
requests==2.28.2
aws-lambda-powertools==2.5.0